pyproject.toml
[project]
name = "conference-tracker"
version = "0.1.0"
description = "Conference deadline tracker for ML/AI and Python conferences"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
scraper = [
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "llama-cpp-python>=0.2.0",
]
